+++ role-guardian/src/main/java/org/rfcx/guardian/guardian/audio/encode/AudioEncodeJobService.java
@@ -113,10 +113,10 @@ public class AudioEncodeJobService extends Service {
 							app.audioEncodeDb.dbEncodeQueue.incrementSingleRowAttempts(latestQueuedAudioToEncode[1]);
 
 							Log.d("encoding","postEncodeFile");
-							File postEncodeFile = new File(RfcxAudioUtils.getAudioFileLocation_PostEncode(context, (long) Long.parseLong(latestQueuedAudioToEncode[1]),latestQueuedAudioToEncode[6]));
+							File postEncodeFile = new File(RfcxAudioUtils.getAudioFileLocation_PostEncode(context, Long.parseLong(latestQueuedAudioToEncode[1]), latestQueuedAudioToEncode[6]));
 
 							Log.d("encoding","Gzipfile");
-							File gZippedFile = new File(RfcxAudioUtils.getAudioFileLocation_Complete_PostGZip(app.rfcxGuardianIdentity.getGuid(), context, (long) Long.parseLong(latestQueuedAudioToEncode[1]),RfcxAudioUtils.getFileExtension(latestQueuedAudioToEncode[6])));
+							File gZippedFile = new File(RfcxAudioUtils.getAudioFileLocation_Complete_PostGZip(app.rfcxGuardianIdentity.getGuid(), context, Long.parseLong(latestQueuedAudioToEncode[1]), RfcxAudioUtils.getFileExtension(latestQueuedAudioToEncode[6])));
 
 							// just in case there's already a post-encoded file, delete it first
 							Log.d("encoding","check if exist");
@@ -131,8 +131,8 @@ public class AudioEncodeJobService extends Service {
 									preEncodeFile, 											// source file
 									postEncodeFile, 											// target file
 									latestQueuedAudioToEncode[6], 							// encoding codec
-									(int) Integer.parseInt(latestQueuedAudioToEncode[4]), 	// encoding sample rate
-									(int) Integer.parseInt(latestQueuedAudioToEncode[5]), 	// encoding target bitrate, if codec-supported
+									Integer.parseInt(latestQueuedAudioToEncode[4]), 	// encoding sample rate
+									Integer.parseInt(latestQueuedAudioToEncode[5]), 	// encoding target bitrate, if codec-supported
 									10											// encoding quality, if codec-supported
 								);
 
@@ -160,10 +160,10 @@ public class AudioEncodeJobService extends Service {
 											latestQueuedAudioToEncode[1], 
 											RfcxAudioUtils.getFileExtension(latestQueuedAudioToEncode[6]), 
 											preZipDigest, 
-											(int) Integer.parseInt(latestQueuedAudioToEncode[4]), 
+											Integer.parseInt(latestQueuedAudioToEncode[4]),
 											encodeBitRate, 
 											latestQueuedAudioToEncode[6], 
-											(long) Long.parseLong(latestQueuedAudioToEncode[7]),
+											Long.parseLong(latestQueuedAudioToEncode[7]),
 											encodeDuration, 
 											gZippedFile.getAbsolutePath()
 										);
